Refactor Home component to utilize useMedia hook for data fetching; remove redundant state management and API calls
