+++ src/hooks/apiHooks.js
@@ -1,0 +1,47 @@
+import {useEffect, useState} from 'react';
+
+import {fetchData} from '../utils/fetchData';
+import {uniqBy} from 'lodash';
+
+const authApiUrl = import.meta.env.VITE_AUTH_API;
+const mediaApiUrl = import.meta.env.VITE_MEDIA_API;
+
+const useMedia = () => {
+  const [mediaArray, setMediaArray] = useState([]);
+
+  const getMedia = async () => {
+    try {
+      const mediaData = await fetchData(`${mediaApiUrl}/media`);
+      const uniqueUserIds = uniqBy(mediaData, 'user_id');
+
+      const userData = await Promise.all(
+        uniqueUserIds.map((item) =>
+          fetchData(`${authApiUrl}/users/${item.user_id}`),
+        ),
+      );
+
+      // duplikaattien poisto on tehtävänannon ulkopuolella, ei tarvitse toteuttaa
+      const userMap = userData.reduce((map, {user_id, username}) => {
+        map[user_id] = username;
+        return map;
+      }, {});
+
+      const newData = mediaData.map((item) => ({
+        ...item,
+        username: userMap[item.user_id],
+      }));
+
+      setMediaArray(newData);
+    } catch (error) {
+      console.error('error', error);
+    }
+  };
+
+  useEffect(() => {
+    getMedia();
+  }, []);
+
+  return mediaArray;
+};
+
+export default useMedia;
